+++ ChangeLog.md
@@ -2,5 +2,23 @@
 
 ## **V0.0.1 - Release**
 Inital Release
-- vvb
-- vfb
+- Added Main Functions:
+    - Get-DiscogsPSArtist
+    - Get-DiscogsPSArtistReleases
+    - Get-DiscogsPSLabel
+    - Get-DiscogsPSLabelReleases
+    - Get-DiscogsPSMaster
+    - Get-DiscogsPSRelease
+    - Get-DiscogsPSReleaseRating
+    - Get-DiscogsPSReleaseRatingByUser
+    - Search-DiscogsPSDatabase
+- Added Classes:
+    - DiscogsArtist
+    - DiscogsArtistMember
+    - DiscogsArtistRef
+- Added Helper Fucntions:
+    - Add-URIArgument
+    - Convert-URIArgument
+    - Invoke-DiscogsPSPaging
+- Added Tests:
+    - Base Module Tests